+++ solver_rust_conic/totsu/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Totally revised in order to support CUDA linear algebra.
+- Divided into 4 crates.
 ### Deprecated
 ### Removed
 ### Fixed